backend/requirements-vercel.txt
# Core FastAPI Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database Dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.2.0

# Astrology Dependencies
pyswisseph==2.10.3.2

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utility Dependencies
pytz==2023.3
python-dateutil==2.8.2
